fixed mobs attack range. implemented a basic way to handle with shield skills. implemented inventory lime limit and warehouse lime limit to 2.100.000.000. fixed the npc shop when buying items.

+++ jreunion/src/main/java/org/reunionemu/jreunion/game/ItemType.java
@@ -19,7 +19,7 @@ public class ItemType{
 
 	private String name;
 	
-	private int maxExtraStats;
+	private long maxExtraStats;
 	
 	private int maxDurability;
 	
@@ -212,11 +212,11 @@ public class ItemType{
 		this.maxDurability = maxDurability;
 	}
 
-	public int getMaxExtraStats() {
+	public long getMaxExtraStats() {
 		return maxExtraStats;
 	}
 
-	public void setMaxExtraStats(int maxExtraStats) {
+	public void setMaxExtraStats(long maxExtraStats) {
 		this.maxExtraStats = maxExtraStats;
 	}
